--- a/crystal/src/crystal.cr
+++ b/crystal/src/crystal.cr
@@ -8,8 +8,7 @@ INITIAL_TAG_ARRAY_CAP = 256
 struct Post
   include JSON::Serializable
 
-  @[JSON::Field(key: "_id")]
-  property id : String
+  property _id : String
   property title : String
   property tags : Array(String)
 end
@@ -17,13 +16,12 @@ end
 struct RelatedPost
   include JSON::Serializable
 
-  def initialize(@id : String, @tags : Array(String), @related : Array(Post))
+  def initialize(@_id : String, @tags : Array(String), @related : Array(Post?))
   end
 
-  @[JSON::Field(key: "_id")]
-  property id : String
+  property _id : String
   property tags : Array(String)
-  property related : Array(Post)
+  property related : Array(Post?)
 end
 
 posts = Array(Post).from_json(File.read("../posts.json"))
@@ -40,7 +38,7 @@ posts.each_with_index do |post, i|
   end
 end
 
-allRelatedPosts = Array(RelatedPost?).new(posts.size, nil)
+allRelatedPosts = Array(RelatedPost?).new(posts.size)
 # using Pointer buffer can greatly improve performance but this is unsafe, thus against the benchmark rules
 # https://crystal-lang.org/reference/latest/syntax_and_semantics/unsafe.html
 tagged_post_count = Array(Int32).new(posts.size, 0) # Pointer(Int32).malloc(posts.size, 0)
@@ -57,36 +55,30 @@ posts.each_with_index do |post, idx|
   tagged_post_count[idx] = 0 # don't count self
 
   # flattened list of (count, id), size at 6 to avoid resizing. also faster than allocating outside loop
-  top5 = Array(Int32).new(TOPN * 2, 0) # Pointer(Int32).malloc(TOPN * 2, 0)
+  top5 = Array(Int32).new(TOPN, 0) # Pointer(Int32).malloc(TOPN * 2, 0)
+  top_posts = Array(Post?).new(TOPN, nil)
   min_tags = 0
 
   posts.size.times do |j|
     count = tagged_post_count[j]
     if count > min_tags
-      upper_bound = (TOPN - 2) * 2
+      upper_bound = (TOPN - 2)
 
       while upper_bound >= 0 && count > top5[upper_bound]
-        top5[upper_bound + 2] = top5[upper_bound]
-        top5[upper_bound + 3] = top5[upper_bound + 1]
-        upper_bound -= 2
+        top5[upper_bound + 1] = top5[upper_bound]
+        top_posts[upper_bound + 1] = top_posts[upper_bound]
+        upper_bound -= 1
       end
 
-      upper_bound += 2
+      upper_bound += 1
       top5[upper_bound] = count
-      top5[upper_bound + 1] = j
+      top_posts[upper_bound] = posts[j]
 
-      min_tags = top5[TOPN*2 - 2]
+      min_tags = top5[TOPN - 1]
     end
   end
 
-  top_posts = Array(Post).new(TOPN)
-
-  # Convert indexes back to Post pointers
-  (1...10).step(2) do |i|
-    top_posts << posts[top5[i]]
-  end
-
-  allRelatedPosts[idx] = RelatedPost.new(id: post.id, tags: post.tags, related: top_posts)
+  allRelatedPosts << RelatedPost.new(_id: post._id, tags: post.tags, related: top_posts)
 end
 
 t2 = Time.monotonic