@jinyus 2023-10-17
Main branch:   

    65.16 ms 
    976.41 ms 
    8.67 s

This PR:

```
Crystal:

        Benchmark 1: ./crystal
        Processing time (w/o IO): 61.235728ms
        Processing time (w/o IO): 60.584318ms
        Processing time (w/o IO): 60.648118ms
        Processing time (w/o IO): 60.465715ms
        Processing time (w/o IO): 60.211711ms
        Processing time (w/o IO): 60.369214ms
        Processing time (w/o IO): 60.489416ms
        Processing time (w/o IO): 60.588317ms
        Processing time (w/o IO): 61.163027ms
        Processing time (w/o IO): 60.939624ms
        Processing time (w/o IO): 60.244812ms
        Processing time (w/o IO): 60.534317ms
        Processing time (w/o IO): 60.13621ms
          Time (mean ± σ):     111.8 ms ±   1.5 ms    [User: 102.3 ms, System: 9.3 ms]
          Range (min … max):   109.6 ms … 113.9 ms    10 runs
         
Crystal:

        Benchmark 1: ./crystal
        Processing time (w/o IO): 911.692512ms
        Processing time (w/o IO): 908.057565ms
        Processing time (w/o IO): 911.618158ms
          Time (mean ± σ):      1.124 s ±  0.004 s    [User: 1.092 s, System: 0.031 s]
          Range (min … max):    1.121 s …  1.126 s    2 runs
         
Crystal:

        Benchmark 1: ./crystal
        Processing time (w/o IO): 8036.191896ms
        Processing time (w/o IO): 8024.284711ms
        Processing time (w/o IO): 8016.027242ms
          Time (mean ± σ):      8.828 s ±  0.009 s    [User: 8.680 s, System: 0.138 s]
          Range (min … max):    8.821 s …  8.834 s    2 runs
```